refactor: Move `ServerReceiveState` into own file
