+++ src/server.rs
@@ -2,7 +2,7 @@ use std::fmt::{Debug, Formatter};
 
 use imap_codec::{
     decode::{AuthenticateDataDecodeError, CommandDecodeError, IdleDoneDecodeError},
-    AuthenticateDataCodec, CommandCodec, GreetingCodec, IdleDoneCodec, ResponseCodec,
+    CommandCodec, GreetingCodec, ResponseCodec,
 };
 use imap_types::{
     auth::AuthenticateData,
@@ -18,6 +18,7 @@ use crate::{
     handle::{Handle, HandleGenerator, HandleGeneratorGenerator, RawHandle},
     receive::{ReceiveError, ReceiveEvent, ReceiveState},
     send_response::{SendResponseEvent, SendResponseState},
+    server_receive::{NextExpectedMessage, ServerReceiveState},
     types::CommandAuthenticate,
     Flow, FlowInterrupt,
 };
@@ -466,73 +467,6 @@ impl ServerFlow {
     }
 }
 
-#[derive(Clone, Copy)]
-enum NextExpectedMessage {
-    Command,
-    AuthenticateData,
-    IdleAccept,
-    IdleDone,
-}
-
-enum ServerReceiveState {
-    Command(ReceiveState<CommandCodec>),
-    AuthenticateData(ReceiveState<AuthenticateDataCodec>),
-    IdleAccept(ReceiveState<NoCodec>),
-    IdleDone(ReceiveState<IdleDoneCodec>),
-    // This state is set only temporarily during `ServerReceiveState::change_state`
-    Dummy,
-}
-
-impl ServerReceiveState {
-    fn change_state(&mut self, next_expected_message: NextExpectedMessage) {
-        // NOTE: This function MUST NOT panic. Otherwise the dummy state will remain indefinitely.
-        let old_state = std::mem::replace(self, ServerReceiveState::Dummy);
-        let new_state = match next_expected_message {
-            NextExpectedMessage::Command => {
-                let codec = CommandCodec::default();
-                Self::Command(match old_state {
-                    Self::Command(state) => state,
-                    Self::AuthenticateData(state) => state.change_codec(codec),
-                    Self::IdleAccept(state) => state.change_codec(codec),
-                    Self::IdleDone(state) => state.change_codec(codec),
-                    Self::Dummy => unreachable!(),
-                })
-            }
-            NextExpectedMessage::AuthenticateData => {
-                let codec = AuthenticateDataCodec::default();
-                Self::AuthenticateData(match old_state {
-                    Self::Command(state) => state.change_codec(codec),
-                    Self::AuthenticateData(state) => state,
-                    Self::IdleAccept(state) => state.change_codec(codec),
-                    Self::IdleDone(state) => state.change_codec(codec),
-                    Self::Dummy => unreachable!(),
-                })
-            }
-            NextExpectedMessage::IdleAccept => {
-                let codec = NoCodec;
-                Self::IdleAccept(match old_state {
-                    Self::Command(state) => state.change_codec(codec),
-                    Self::AuthenticateData(state) => state.change_codec(codec),
-                    Self::IdleAccept(state) => state,
-                    Self::IdleDone(state) => state.change_codec(codec),
-                    Self::Dummy => unreachable!(),
-                })
-            }
-            NextExpectedMessage::IdleDone => {
-                let codec = IdleDoneCodec::default();
-                Self::IdleDone(match old_state {
-                    Self::Command(state) => state.change_codec(codec),
-                    Self::AuthenticateData(state) => state.change_codec(codec),
-                    Self::IdleAccept(state) => state.change_codec(codec),
-                    Self::IdleDone(state) => state,
-                    Self::Dummy => unreachable!(),
-                })
-            }
-        };
-        *self = new_state;
-    }
-}
-
 /// A handle for an enqueued [`Response`].
 ///
 /// This handle can be used to track the sending progress. After a [`Response`] was enqueued via
@@ -614,6 +548,3 @@ pub enum ServerFlowError {
     #[error("Command was rejected because it was too long")]
     CommandTooLong { discarded_bytes: Secret<Box<[u8]>> },
 }
-
-/// A dummy codec we use for technical reasons when we don't want to receive anything at all.
-struct NoCodec;
